changed data type in gcd optimized algo and added lcm optimized program using euclid's algo.

+++ src/mathematics/Gcd_Euclied_Algo_Optimized.java
@@ -11,7 +11,7 @@ public class Gcd_Euclied_Algo_Optimized {
         System.out.println(gcd(100000000, 200000000));
     }
 
-    static int gcd(int a, int b) {
+    static long gcd(long a, long b) {
         if (b == 0) return a;
 
         return gcd(b, a % b);
